Split repository.ts: extract the moderation domain (7/N)
Seventh slice of the repository.ts carve-up (audit L14). Moves three sections
into repository/moderation.ts, kept in their original order with their section
banners intact:

  auto-moderation strikes, the bot-side WhatsApp block list (#572), and the
  durable appeal_moderation record (#554).

One module rather than three: they are three sections of the same domain, and
the appeals section is separated from the other two in repository.ts only by
unrelated intervening sections.

None of these carries a conversation-scoped admin read, so the 🔒 SQL-scoping
invariant is not in play. The scoped reporting/feedback reads deliberately stay
in repository.ts for their own dedicated extraction, where the scoping can be
reviewed on its own rather than beside routine moves.

All 21 moved symbols verified byte-identical against origin/main.
repository.ts is now 5,362 lines, down from 7,137 at the start of the campaign.
